Rename solve module to 'logic'
