+++ gaslines/display.py
@@ -6,7 +6,7 @@ solutions in a standard terminal.
 
 import time
 
-from gaslines.solve import has_head
+from gaslines.logic import has_head
 from gaslines.utility import get_number_of_rows
